requirements.txt
pyyaml
seaborn
matplotlib
gym